# api-first/openapi-inditex.yaml
openapi: 3.0.0
info:
  title: SGA - Inditex
  description: Especificación de API para prueba técnica SGA-Inditex.
  version: 1.0.0
  contact:
    email: storres@paradigmadigital.com

servers:
  - url: http://localhost/

tags:
  - name: product
    description: Búsqueda de productos
    externalDocs:
      description: Encuentra más sobre nuestros productos
      url: https://www.inditex.com/es/home

paths:
  /product:
    get:
      tags:
        - product
      summary: Búsqueda producto por fecha
      description: Producto encontrado entre una fecha de entrada y una salida con precio correspondiente
      operationId: findById
      parameters:
        - name: entryDate
          in: query
          description: Fecha de aplicación del producto
          required: true
          schema:
            type: string
            format: date-time
        - name: productId
          in: query
          description: Identificador del producto
          required: true
          schema:
            type: integer
        - name: brandId
          in: query
          description: Identificador de la cadena
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '400':
          description: Petición errónea
        '404':
          description: Producto no encontrado
        '500':
          description: Error del servidor. Contacte con soporte

components:
  schemas:
    Price:
      type: object
      properties:
        productId:
          type: integer
          description: Identificador código de producto
        startDate:
          type: string
          format: date-time
          example: "2020-06-16T21:00:00"
        endDate:
          type: string
          format: date-time
          example: "2020-06-16T21:00:00"
        priceList:
          type: integer
        price:
          type: number
          format: double
        curr:
          type: string
          description: Currency. Podría convertirse en enum si se incluyen más
          default: EUR
        brandId:
          type: integer
          description: Identificador código de cadena
        priority:
          type: integer
          description: Desambiguador de aplicación de precios

    Product:
      type: object
      properties:
        productId:
          type: integer
          description: Identificador código de producto
        startDate:
          type: string
          format: date-time
          example: "2020-06-16T21:00:00"
        endDate:
          type: string
          format: date-time
          example: "2020-06-16T21:00:00"
        price:
          type: number
          format: double
        curr:
          type: string
          default: EUR
          description: Currency. Podría convertirse en enum si se incluyen más
        brandId:
          type: integer
          description: Identificador código de cadena
        priority:
          type: integer
          description: Desambiguador de aplicación de precios

    Brand:
      type: object
      properties:
        id:
          type: integer
          description: Identificador código de producto
        groupName:
          type: string
          description: Nombre del grupo (ZARA)

externalDocs:
  description: Find out more about Inditex
  url: https://www.inditex.com/es/home